+++ src/main/java/spring/dataBase/repository/entity/User.java
@@ -15,11 +15,11 @@ import java.util.List;
 @Builder
 @Entity
 @Table(name = "users")
-public class User implements BaseEntity<Long> {
+public class User implements BaseEntity<Integer> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false, unique = true)
     private String username;
